src/main/java/com/leandrosouza/leandroagrotis/config/ConfigBeans.java
package com.leandrosouza.leandroagrotis.config;

import org.springframework.context.MessageSource;
import org.springframework.context.annotation.Bean;
import org.springframework.context.annotation.Configuration;
import org.springframework.context.support.ReloadableResourceBundleMessageSource;

import java.util.Locale;

@Configuration
public class ConfigBeans {

    @Bean
    public MessageSource messageSource() {
        ReloadableResourceBundleMessageSource messageSource = new ReloadableResourceBundleMessageSource();
        messageSource.setBasename("classpath:messages");
        messageSource.setCacheSeconds(300);
        return messageSource;
    }

    @Bean
    public Locale locale() {
        return new Locale("pt", "BR");
    }
}
